--- src/VisualStudio/Core/Def/PublicAPI.Unshipped.txt
@@ -1,1 +1,0 @@
-
